Create array
In this code we have made an array of 10 integers. Then we take input from the user to find the number in the array. If the number is present in the array we print " the number is present in the array"
else we print" the number is not present in the array".